.hpp->.h, .tpp->.cpp

+++ main.cpp
@@ -5,7 +5,7 @@
 #include <functional>
 
 #include "Particle.h"
-#include "Quadtree.hpp"
+#include "Quadtree.h"
 
 using namespace std;
 
@@ -14,8 +14,8 @@ bool particleCollision(const Particle& particle, const Rectangle& range) {
 }
 
 int main() {
-    unsigned int screenWidth = 1920; 
-    unsigned int screenHeight = 1080; 
+    float screenWidth = 1920; 
+    float screenHeight = 1080; 
 
     InitWindow(screenWidth, screenHeight, "DLA");
     //SetTargetFPS(100);
